chore(node/service): Squash service traits
